pi0ninja_v3/pyproject.toml
[project]
name = "pi0ninja-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow",
    "pi0disp",
    "piservo0",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "google-generativeai",
    "python-dotenv",
    "websockets",
    "googlesearch-python",
    "pyngrok",
    "python-multipart",
    "qrcode[pil]",
]

[project.scripts]
web-server = "pi0ninja_v3.web_server:main"

[project.optional-dependencies]
dev = [
    "ruff",
]


[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
